기존 NetworkingManager의 이름을 APINetworkingManager로 바꾼 후, NetworkingManager 싱글톤 객체 추가 구현
